// src/components/Types/color-type.enum.ts
export enum ColorType {
    Default = 'rgba(107, 108, 111, 100)',
    Level1 = 'rgba(151, 97, 83, 100)',
    Level2 = 'rgba(177, 131, 64, 100)',
    Level3 = 'rgba(199, 172, 94, 100)',
    Level4 = 'rgba(150, 146, 77, 100)',
    Level5 = 'rgba(95, 117, 100, 100)',
    Level6 = 'rgba(115, 154, 153, 100)',
    Level7 = 'rgba(135, 161, 187, 100)',
    Level8 = 'rgba(143, 137, 194, 100)',
    Level9 = 'rgba(164, 129, 186, 100)',
    Level10 = 'rgba(199, 126, 152, 100)',
}